[lazy arena] phase in new arena based predicate pushdown optimizer
