--- polars/polars-lazy/src/future/logical_plan/mod.rs
@@ -1,5 +1,0 @@
-mod arena;
-pub mod optimizer;
-use crate::logical_plan::Context;
-use crate::prelude::*;
-use polars_core::prelude::*;
