--- polars/polars-lazy/src/future/logical_plan/optimizer/mod.rs
@@ -1,2 +1,0 @@
-#[cfg(feature = "future")]
-pub(crate) mod predicate_pushdown;
