+++ polars/polars-lazy/src/logical_plan/optimizer/mod.rs
@@ -1,11 +1,9 @@
-use std::collections::HashMap;
-
 use ahash::RandomState;
-
 use polars_core::frame::group_by::{fmt_groupby_column, GroupByMethod};
 use polars_core::frame::hash_join::JoinType;
 use polars_core::prelude::*;
 use polars_core::utils::{get_supertype, Arena, Node};
+use std::collections::HashMap;
 
 use crate::logical_plan::{prepare_projection, Context};
 use crate::prelude::*;
@@ -13,7 +11,7 @@ use crate::utils::{expr_to_root_column_exprs, rename_field};
 
 pub(crate) mod aggregate_pushdown;
 pub(crate) mod aggregate_scan_projections;
-pub(crate) mod predicate;
+pub(crate) mod predicate_pushdown;
 pub(crate) mod projection;
 pub(crate) mod simplify_expr;
 pub(crate) mod type_coercion;
